Log type and response body that has failed to deserialize
